Share CreatureFlags
