+++ SharedSE/SharedSE.vcxitems.filters
@@ -372,6 +372,9 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3DialogueFlags.h">
       <Filter>Header Files\TES3</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3CreatureFlags.h">
+      <Filter>Header Files\TES3</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
